requirements.txt
ctgan==0.4.0
sdv==0.8.0
table_evaluator==1.2.2.post1
pandas==1.1.4
matplotlib==3.3.2
streamlit==0.75.0
numpy==1.20.1
Pillow==8.2.0
